solution/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Advertisting Service REST API"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=13.4.0",
    "aiogram>=3.18.0",
    "aiohttp>=3.11.11",
    "alembic>=1.14.0",
    "dishka[fastapi]>=1.4.2",
    "fastapi[standard]>=0.115.6",
    "psycopg[binary]>=3.2.4",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.uv]
dev-dependencies = [
    "isort>=6.0.0",
    "ruff>=0.9.2",
    "pytest-asyncio>=0.15.3",
    "pytest==7.2",
    "tavern[pytest]>=2.11.0",
]

[tool.pytest.ini_options]
tavern-global-cfg = "app/tests/tavern/config.yml"
tavern-strict = "json:off headers:off"

log_cli = true
log_cli_level = "INFO"

filterwarnings = "ignore::DeprecationWarning"
